pyproject.toml
[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'

[project]
name = 'captiq'
version = '1.0.0'
description = 'UK Capital Gains Tax Calculator'
readme = 'README.md'
authors = [
    {name = 'Fraser Love', email = 'me@fraser.love'}
]
dependencies = [
    'click~=8.1.0',
    'py-moneyed~=3.0',
    'platformdirs~=4.2.0',
    'prettytable~=3.10.0',
    'python-dateutil~=2.9.0',
    'PyYAML~=6.0.0',
    'typer~=0.12.0',
    'yfinance~=0.2.40'
]
requires-python = '>=3.10'

[project.scripts]
captiq = 'captiq.cli:main'